chore: upgrade Spring Boot to 3.2.x
Update Spring Boot to 3.2.1, the following are required to support
migration:
 - Mongock driver v3 replacing with v4
 - Removed `javax` imports replaced with `jakarta` imports
 - AWS Spring Cloud requires version 3.x.x

Update the AWS SQS config to use the new SQS starter from version
`3.x.x`, instead of the old generic messaging starter.
The new version has better auto-configuration, so the existing SQS
configuration can be deleted.
The `@SqsListener` annotation no longer has a deletion policy, the
default matches the old value so no further configuration is needed.

TIS21-SHED
