--- src/main/java/uk/nhs/hee/trainee/details/config/AmazonSqsConfig.java
@@ -1,73 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright 2023 Crown Copyright (Health Education England)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
- * associated documentation files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
- * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package uk.nhs.hee.trainee.details.config;
-
-import com.amazonaws.services.sqs.AmazonSQSAsync;
-import com.amazonaws.services.sqs.AmazonSQSAsyncClientBuilder;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.awspring.cloud.messaging.core.QueueMessagingTemplate;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.messaging.converter.MappingJackson2MessageConverter;
-import org.springframework.messaging.converter.MessageConverter;
-
-/**
- * Configuration for SQS functionality.
- */
-@Configuration
-public class AmazonSqsConfig {
-
-  /**
-   * Create a default {@link AmazonSQSAsync} bean.
-   *
-   * @return The created bean.
-   */
-  @Bean
-  @Primary
-  public AmazonSQSAsync amazonSqsAsync() {
-    return AmazonSQSAsyncClientBuilder.defaultClient();
-  }
-
-  /**
-   * Create a default {@link QueueMessagingTemplate} bean.
-   *
-   * @return The created bean.
-   */
-  @Bean
-  public QueueMessagingTemplate queueMessagingTemplate() {
-    return new QueueMessagingTemplate(amazonSqsAsync());
-  }
-
-  /**
-   * Create a message converter bean with an object mapper.
-   *
-   * @param objectMapper The object mapper to set.
-   * @return The created message converter.
-   */
-  @Bean
-  public MessageConverter messageConverter(ObjectMapper objectMapper) {
-    var converter = new MappingJackson2MessageConverter();
-    converter.setObjectMapper(objectMapper);
-    return converter;
-  }
-}
