+++ src/main/java/uk/nhs/hee/trainee/details/config/ApplicationConfiguration.java
@@ -21,9 +21,12 @@
 
 package uk.nhs.hee.trainee.details.config;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.messaging.converter.MappingJackson2MessageConverter;
+import org.springframework.messaging.converter.MessageConverter;
 import org.springframework.web.client.RestTemplate;
 
 /**
@@ -35,5 +38,18 @@ public class ApplicationConfiguration {
   @Bean
   RestTemplate restTemplate(RestTemplateBuilder builder) {
     return builder.build();
+  }
+
+  /**
+   * Create a message converter bean with an object mapper.
+   *
+   * @param objectMapper The object mapper to set.
+   * @return The created message converter.
+   */
+  @Bean
+  public MessageConverter messageConverter(ObjectMapper objectMapper) {
+    var converter = new MappingJackson2MessageConverter();
+    converter.setObjectMapper(objectMapper);
+    return converter;
   }
 }
