+++ src/main/java/uk/nhs/hee/trainee/details/config/AwsXrayConfiguration.java
@@ -21,8 +21,8 @@
 
 package uk.nhs.hee.trainee.details.config;
 
-import com.amazonaws.xray.javax.servlet.AWSXRayServletFilter;
-import javax.servlet.Filter;
+import com.amazonaws.xray.jakarta.servlet.AWSXRayServletFilter;
+import jakarta.servlet.Filter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
