+++ src/main/java/uk/nhs/hee/trainee/details/config/MongoConfiguration.java
@@ -21,7 +21,7 @@
 
 package uk.nhs.hee.trainee.details.config;
 
-import javax.annotation.PostConstruct;
+import jakarta.annotation.PostConstruct;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.mongodb.core.MongoTemplate;
