+++ src/main/java/uk/nhs/hee/trainee/details/config/SentryConfiguration.java
@@ -22,8 +22,8 @@
 package uk.nhs.hee.trainee.details.config;
 
 import io.sentry.Sentry;
+import jakarta.annotation.PostConstruct;
 import java.util.Optional;
-import javax.annotation.PostConstruct;
 import org.springframework.context.annotation.Configuration;
 import uk.nhs.hee.trainee.details.config.EcsMetadataConfiguration.EcsMetadata;
