+++ src/main/java/uk/nhs/hee/trainee/details/dto/PersonalDetailsUpdateEvent.java
@@ -22,8 +22,8 @@
 package uk.nhs.hee.trainee.details.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 
 /**
  * A representation of a PersonalDetails update event.
