+++ src/main/resources/application.yml
@@ -4,15 +4,19 @@ server:
     context-path: /trainee
 
 spring:
+  cloud:
+    aws:
+      endpoint: ${AWS_ENDPOINT:}
   data:
     mongodb:
       uri: mongodb://${DB_USER:admin}:${DB_PASSWORD:pwd}@${DB_HOST:localhost}:${DB_PORT:27017}/${DB_NAME:trainee}?authSource=admin
-  redis:
-    host: ${REDIS_HOST:localhost}
-    port: ${REDIS_PORT:6379}
-    ssl: ${REDIS_SSL:false}
-    user: ${REDIS_USER:default}
-    password: ${REDIS_PASSWORD:password}
+    redis:
+      host: ${REDIS_HOST:localhost}
+      port: ${REDIS_PORT:6379}
+      ssl:
+        enabled: ${REDIS_SSL:false}
+      user: ${REDIS_USER:default}
+      password: ${REDIS_PASSWORD:password}
   rabbitmq:
     host: ${TRAINEE_DETAILS_RABBITMQ_HOST}
     port: ${TRAINEE_DETAILS_RABBITMQ_PORT}
