--- src/test/java/uk/nhs/hee/trainee/details/config/AmazonSqsConfigTest.java
@@ -1,55 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright 2023 Crown Copyright (Health Education England)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
- * associated documentation files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
- * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package uk.nhs.hee.trainee.details.config;
-
-import static org.hamcrest.CoreMatchers.instanceOf;
-import static org.hamcrest.CoreMatchers.sameInstance;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.messaging.converter.MappingJackson2MessageConverter;
-import org.springframework.messaging.converter.MessageConverter;
-
-class AmazonSqsConfigTest {
-
-  private AmazonSqsConfig config;
-
-  @BeforeEach
-  void setUp() {
-    config = new AmazonSqsConfig();
-  }
-
-  @Test
-  void shouldIncludeExistingObjectMapperInMessageConverter() {
-    ObjectMapper objectMapper = new ObjectMapper();
-
-    MessageConverter messageConverter = config.messageConverter(objectMapper);
-
-    assertThat("Unexpected message converter.", messageConverter,
-        instanceOf(MappingJackson2MessageConverter.class));
-    assertThat("Unexpected object mapper.",
-        ((MappingJackson2MessageConverter) messageConverter).getObjectMapper(),
-        sameInstance(objectMapper));
-  }
-}
