+++ src/test/resources/application-test.yml
@@ -4,4 +4,6 @@ mongock:
 cloud:
   aws:
     region:
-      static: eu-west-2
+      static: aws-global
+    sqs:
+      enabled: false
